refactor: inject GOOS into security analyzers
