+++ internal/runner/security/file_validation_symlink_darwin.go
@@ -1,0 +1,27 @@
+//go:build darwin
+
+package security
+
+import "path/filepath"
+
+// isAllowedOSManagedSymlink allows known macOS root-level aliases only when
+// their resolved target exactly matches the expected destination.
+func isAllowedOSManagedSymlink(path string) bool {
+	allowedTargets := map[string]string{
+		"/tmp": "/private/tmp",
+		"/var": "/private/var",
+	}
+
+	cleanPath := filepath.Clean(path)
+	expectedTarget, ok := allowedTargets[cleanPath]
+	if !ok {
+		return false
+	}
+
+	resolvedPath, err := filepath.EvalSymlinks(cleanPath)
+	if err != nil {
+		return false
+	}
+
+	return filepath.Clean(resolvedPath) == expectedTarget
+}
